+++ R/provider-portkey.R
@@ -12,22 +12,21 @@
 #' @family chatbots
 #' @param api_key `r api_key_param("PORTKEY_API_KEY")`
 #' @param model `r param_model("gpt-4o", "openai")`
-#' @param virtual_key A virtual identifier storing LLM provider's API key. See
-#'   [documentation](https://portkey.ai/docs/product/ai-gateway/virtual-keys).
-#'   Can be read from the `PORTKEY_VIRTUAL_KEY` environment variable.
+#' @param virtual_key `r lifecycle::badge("deprecated")` No longer supported by Portkey.
+#' Read more: https://portkey.ai/docs/support/upgrade-to-model-catalog
 #' @export
 #' @inheritParams chat_openai
 #' @inherit chat_openai return
 #' @examples
 #' \dontrun{
-#' chat <- chat_portkey(virtual_key = Sys.getenv("PORTKEY_VIRTUAL_KEY"))
+#' chat <- chat_portkey()
 #' chat$chat("Tell me three jokes about statisticians")
 #' }
 chat_portkey <- function(
   system_prompt = NULL,
   base_url = "https://api.portkey.ai/v1",
   api_key = portkey_key(),
-  virtual_key = portkey_virtual_key(),
+  virtual_key = deprecated(),
   model = NULL,
   params = NULL,
   api_args = list(),
@@ -36,7 +35,13 @@ chat_portkey <- function(
 ) {
   model <- set_default(model, "gpt-4o")
   echo <- check_echo(echo)
-
+  if (lifecycle::is_present(virtual_key)) {
+    lifecycle::deprecate_warn(
+      when = "0.4.0",
+      what = "chat_portkey(virtual_key=)",
+      with = "chat_portkey()",
+    )
+  }
   params <- params %||% params()
   provider <- ProviderPortkeyAI(
     name = "PortkeyAI",
@@ -45,7 +50,6 @@ chat_portkey <- function(
     params = params,
     extra_args = api_args,
     api_key = api_key,
-    virtual_key = virtual_key,
     extra_headers = api_headers
   )
   Chat$new(provider = provider, system_prompt = system_prompt, echo = echo)
@@ -65,10 +69,7 @@ chat_portkey_test <- function(
 
 ProviderPortkeyAI <- new_class(
   "ProviderPortkeyAI",
-  parent = ProviderOpenAI,
-  properties = list(
-    virtual_key = prop_string(allow_null = TRUE)
-  )
+  parent = ProviderOpenAI
 )
 
 portkey_key_exists <- function() {
@@ -79,21 +80,11 @@ portkey_key <- function() {
   key_get("PORTKEY_API_KEY")
 }
 
-portkey_virtual_key <- function() {
-  val <- Sys.getenv("PORTKEY_VIRTUAL_KEY")
-  if (!identical(val, "")) {
-    val
-  } else {
-    NULL
-  }
-}
-
 method(base_request, ProviderPortkeyAI) <- function(provider) {
   req <- request(provider@base_url)
   req <- httr2::req_headers(
     req,
-    `x-portkey-api-key` = provider@api_key,
-    `x-portkey-virtual-key` = provider@virtual_key
+    `x-portkey-api-key` = provider@api_key
   )
   req <- ellmer_req_robustify(req)
   req <- ellmer_req_user_agent(req)
@@ -107,14 +98,21 @@ method(base_request, ProviderPortkeyAI) <- function(provider) {
 models_portkey <- function(
   base_url = "https://api.portkey.ai/v1",
   api_key = portkey_key(),
-  virtual_key = NULL
+  virtual_key = deprecated()
 ) {
+  if (lifecycle::is_present(virtual_key)) {
+    lifecycle::deprecate_warn(
+      when = "0.4.0",
+      what = "models_portkey(virtual_key=)",
+      with = "models_portkey()",
+    )
+  }
+  
   provider <- ProviderPortkeyAI(
     name = "PortkeyAI",
     model = "",
     base_url = base_url,
-    api_key = api_key,
-    virtual_key = virtual_key
+    api_key = api_key
   )
 
   req <- base_request(provider)